feat: renaming annotations and support httpMethods
