+++ jmix-rest/rest/src/main/java/io/jmix/rest/annotation/RestHttpMethod.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Haulmont.
+ * Copyright 2024 Haulmont.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,11 +16,7 @@
 
 package io.jmix.rest.annotation;
 
-import java.lang.annotation.*;
-
-@Target({ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface OpenMethod {
-    String value() default "";
+public enum RestHttpMethod {
+    GET,
+    POST,
 }
